Move helpers for message phrasing to own file
