+++ src/jsMain/kotlin/ConfigurationCacheReportPage.kt
@@ -14,7 +14,11 @@
  * limitations under the License.
  */
 
+import data.found
+import data.itsOrTheir
 import data.mapAt
+import data.sIfPlural
+import data.wasOrWere
 import elmish.Component
 import elmish.View
 import elmish.a
@@ -302,27 +306,6 @@ object ConfigurationCacheReportPage :
             if (totalProblems > reportedProblems) "$it, only the first $reportedProblems ${wasOrWere(reportedProblems)} included in this report"
             else it
         }
-
-    private
-    fun found(count: Int, what: String) =
-        "${count.toStringOrNo()} ${what.sIfPlural(count)} ${wasOrWere(count)} found"
-
-    private
-    fun Int.toStringOrNo() =
-        if (this != 0) toString()
-        else "No"
-
-    private
-    fun String.sIfPlural(count: Int) =
-        if (count < 2) this else "${this}s"
-
-    private
-    fun wasOrWere(count: Int) =
-        if (count <= 1) "was" else "were"
-
-    private
-    fun itsOrTheir(count: Int) =
-        if (count <= 1) "its" else "their"
 
     private
     fun displayTabButton(tab: Tab, activeTab: Tab, problemsCount: Int): View<Intent> = div(
